미사용/requirements.txt
gspread>=5.0.0
google-auth>=2.0.0
